+++ inc/instruction.h
@@ -49,7 +49,7 @@ struct ooo_model_instr {
   bool branch_prediction = 0;
   bool branch_mispredicted = 0; // A branch can be mispredicted even if the direction prediction is correct when the predicted target is not correct
 
-  uint16_t asid;
+  int asid;
 
   uint8_t branch_type = NOT_BRANCH;
   uint64_t branch_target = 0;
@@ -74,7 +74,7 @@ struct ooo_model_instr {
 
 private:
   template <typename T>
-  ooo_model_instr(T instr, uint16_t address_space) : ip(instr.ip), is_branch(instr.is_branch), branch_taken(instr.branch_taken), asid(address_space)
+  ooo_model_instr(T instr, int address_space) : ip(instr.ip), is_branch(instr.is_branch), branch_taken(instr.branch_taken), asid(address_space)
   {
     std::remove_copy(std::begin(instr.destination_registers), std::end(instr.destination_registers), std::back_inserter(this->destination_registers), 0);
     std::remove_copy(std::begin(instr.source_registers), std::end(instr.source_registers), std::back_inserter(this->source_registers), 0);
@@ -84,7 +84,7 @@ private:
 
 public:
   explicit ooo_model_instr(input_instr instr) : ooo_model_instr(instr, std::numeric_limits<uint16_t>::max()) {}
-  explicit ooo_model_instr(cloudsuite_instr instr) : ooo_model_instr(instr, (static_cast<uint16_t>(instr.asid[1]) << 8) + instr.asid[0]) {}
+  explicit ooo_model_instr(cloudsuite_instr instr) : ooo_model_instr(instr, ((instr.asid[1]) << 8) + instr.asid[0]) {}
 
   std::size_t num_mem_ops() const { return std::size(destination_memory) + std::size(source_memory); }
